ANDROID: REVERTME: DEBUG: pkvm: x86: bypass legacy IOMMU
This is ONLY for development purpose to allow pkvm able to run on the
platform which only have legacy mode IOMMU.

Introduce a parameter to let pkvm to bypass the enabling of the legacy
IOMMU. With this, although the legacy IOMMU MMIO accessing from host
still traps to pkvm but pkvm will directly forward the accessing to the
IOMMU hardware. So there will be NO IO protections.

To use this, add below in kernel command line: "pkvm=legacy_iommu"

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: 349990461
Test: Android can run protected VMs

Change-Id: I5465603d388d3ddb85eeccdc292646f47b513847
Signed-off-by: Dmytro Maluka <dmaluka@google.com>