ANDROID: pkvm: x86: simplify invept emulation
With using hypercall to do the remote tlb flushing when KVM high has
modified the virtual EPT, KVM high executs the invept instruction only
when it requires to flush EPT TLB in the hardware. So the invept
emulation from pkvm side can be simplified as executing the invept with
desired shadow eptp.

When a VCPU has loaded a shadow vcpu, executing invept with a single
context needs to match with the shadow eptp of this loaded shadow vcpu.
And for global context, directly do invept with the shadow eptp of the
loaded shadow vcpu as there is no other shadow ept's TLB entries left on.

And if a VCPU doesn't have shadow vcpu loaded, its shadow EPT TLB is
already flushed when doing vmclear thus invept is not necessary. So
invept will be ignored for this case.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>

dmaluka@ Without this patch, ARCVM works horribly slow (as the host
issues INVEPT pretty often).

Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: 349990461
Test: Android can run protected VMs

Change-Id: I3a36cfae365956f84423c6a0b3245dfbd549885c
Signed-off-by: Dmytro Maluka <dmaluka@google.com>