ANDROID: pkvm: x86: Add two guest hypercalls to PV MMIO operations
For protected VM, it can't directly access the MMIO region due to it
will exit to the KVM, and KVM will do instruction decode which will
fetch the guest memory. But the guest memory can't be accessed by the
primary VM, so KVM can't emulate the MMIO operation, so MMIO will
failed.

To deal with problem, we can paravirtualize the MMIO operation in
protected VM. The implementaion is to replace all of the MMIO operation
into a hypercall, so it will mitigate the instruction decode in the KVM.
The hypercall will directly tell the KVM the gpa, size, write, value, so
KVM can directly use them to do the MMIO operation, and return the
result to protected VM. This can enable MMIO in protected VM.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>

Bug: 349990461
Test: Android can run protected VMs

Change-Id: I5715cb7f90247fd8e2f9828074276501ecb85814
Signed-off-by: Dmytro Maluka <dmaluka@google.com>