ANDROID: KVM: stats: Add VM stat for remote tlb flush with range
Add a new stat to count the number of kvm_flush_remote_tlbs_with_range()
calls which flushes remote a given remote TLB range.  This is different
from the stat remote_tlb_flush which counts the number of flushing all
remote TLBs.

Flushing remote TLBs could be expensive if KVM is running on a VM, and
adding this new stat may be helpful for nested performance tuning.

Signed-off-by: Zide Chen <zide.chen@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>

Bug: 349990461
Test: Android can run protected VMs

Change-Id: Ie05f698f7f37ea45d540634d01e736c3cf22cc0c
Signed-off-by: Dmytro Maluka <dmaluka@google.com>