ANDROID: pkvm: x86: Fix a minor calculating issue
check_and_init_iommu() tries to get a common set of page level and page
size mask between ept and dmar units. Both page level and size are with
mask value. The current code used min() to calculate mask value which is
not correct. For example, if EPT only supports 5-level while dmar
supports 4-level, the current logic will get the result of 4-level but
the correct result should be 0. This may not happen in reality, though,
as the 5-level capability most likely implies 4-level as well. Still
have the small fix to make the code looking more right.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>

Bug: 349990461
Test: Android can run protected VMs

Change-Id: Ibdf68da96714f89f73c4a230bde12012a74c5bfa
Signed-off-by: Dmytro Maluka <dmaluka@google.com>