ANDROID: pkvm: x86: Dynamically handle host MMIO EPT violation
The host EPT is prepopulated for all the memory and MMIO in the low
address range, but not for the MMIO in high address range. If host VM is
accessing such MMIO, EPT violation happen and pKVM shall do the map
for it if the MMIO range belongs to the host VM, which means this MMIO
cannot be pKVM owned devices(eg. IOMMU) or any pass-through device to
the protected VM (this part will be supported in the future)

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>

Bug: 349990461
Test: "dmesg | grep pkvm" shows "all cpus are in guest mode!"

Change-Id: Ie8041d5f22c62744abfe70ba72ac063ecc21d88c
Signed-off-by: Dmytro Maluka <dmaluka@google.com>