CHROMIUM: mm/mglru: age at different rates
Allow anon and file pages to age at different rates. To prevent a
generation of one type from sticking around forever, the feedback loop
comparison is updated so that a generation of a particular type
accumulates an error cost over time. This is done by tracking the number
of pages evicted to preserve pages in that generation that then later
get refaulted.

BUG=b:342531825
UPSTREAM-TASK=b:357764239
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: Id4078a4a42337fdedf33db5810b031ca45b8abea
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5894973
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Signed-off-by: Hubert Mazur <hmazur@google.com>
Signed-off-by: Tomasz Nowicki <tnowicki@google.com>