CHROMIUM: Add a config to tie generation age together
Add an set-by-default bit to the /sys/kernel/mm/lru_gen/enabled bitmask
to cause new file and anon generations to be created in lockstep. This
uses the same logic to decide when to age and which type of page to
evict as the previous MGLRU aging implementation, so it should replicate
the old behavior.

There are a few places where the previous implementation would compare
file and anon min_seq values. Since the legacy behavior may be enabled
when there is already skew between the file and anon generation numbers,
those comparisons are now done by comparing the number of generations of
each type.

BUG=b:342531825
UPSTREAM-TASK=b:357764239
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: Icbb8e6b588802b68ddb9a5910c1eac1b753bedcf
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5894974
Reviewed-by: Suleiman Souhlal <suleiman@google.com>
Signed-off-by: Hubert Mazur <hmazur@google.com>
Signed-off-by: Tomasz Nowicki <tnowicki@google.com>