ANDROID: GKI: x86: Remove dma_ops from ABI symbols
The upstream patch [1], which is merged into the android15-6.6-desktop
since Ibf06c81f2381cc6dfada71c11938112ba398e34d, removes the dma_ops
symbol from the HDA Intel vendor module.

The symbol list is updated with the following commands:

$ truncate -s0 common/android/abi_gki_x86_64_desktop
$ tools/bazel run \
    //private/google-modules/linux/brya:brya_device_x86_64_abi_update_symbol_list
$ tools/bazel run //common:kernel_x86_64_abi_update

NOTE: The x86 ABI is under construction and it is expected to have
incompatible updates on the android15-6.6-desktop branch.

[1] https://patch.msgid.link/20241002155948.4859-1-tiwai@suse.de

Bug: 360800854
Change-Id: I5d10e742f24efd23a1b4de2b01a9c46a7bd35fdf
Signed-off-by: Slawomir Rosek <srosek@google.com>