ANDROID: KVM: Support FF-A handling for pVMs using the hvc conduit
When the guest VMs are presented with the hvc conduit from the device
tree, the ARM FF-A driver chooses to use this over the smc. This
restricts our ability to trap the FF-A calls from the VMs using the
hvc conduit.

Handle protected VM FF-A calls using the hvc conduit and dispatch the
call to the appropriate handler.

Bug: 269285339
Bug: 278749606
Change-Id: I21bf92ee6ed2dcb175b2b8da3f163c2d2915eb83
Signed-off-by: Jay Monkman <jmonkman@google.com>